feat(check): use branchless logic

+++ src/list/size.c
@@ -6,15 +6,15 @@
 /*   By: smamalig <smamalig@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/12 01:33:28 by smamalig          #+#    #+#             */
-/*   Updated: 2025/02/22 22:10:06 by smamalig         ###   ########.fr       */
+/*   Updated: 2025/02/27 12:26:50 by smamalig         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_list_size(t_list *list)
+size_t	ft_list_size(t_list *list)
 {
-	int	size;
+	size_t	size;
 
 	size = 0;
 	while (list)
